requirements.txt
boto3
botocore
certifi
charset-normalizer
idna
jmespath
numpy
pandas
python-dateutil
python-dotenv
pytz
requests
s3transfer
six
slugify
tzdata
urllib3 